5491-split DictionariesConsistencySetupTest for tests after setup
