+++ tests/php-unit-tests/integration-tests/DictionariesConsistencyAfterSetupTest.php
@@ -1,0 +1,61 @@
+<?php
+/**
+ * Copyright (C) 2013-2020 Combodo SARL
+ * This file is part of iTop.
+ * iTop is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * iTop is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ * You should have received a copy of the GNU Affero General Public License
+ */
+
+namespace Combodo\iTop\Test\UnitTest\Integration;
+
+use Combodo\iTop\Test\UnitTest\ItopTestCase;
+
+class DictionariesConsistencyAfterSetupTest extends ItopTestCase
+{
+	public function FormatProvider(){
+		return [
+			'key does not exist in dictionnary' => [
+				'sTemplate' => null,
+				'sExpectedTraduction' => 'ITOP::DICT:FORMAT:BROKEN:KEY - 1',
+			],
+			'traduction that breaks expected nb of arguments' => [
+				'sTemplate' => 'toto %1$s titi %2$s',
+				'sExpectedTraduction' => 'ITOP::DICT:FORMAT:BROKEN:KEY - 1',
+			],
+			'traduction ok' => [
+				'sTemplate' => 'toto %1$s titi',
+				'sExpectedTraduction' => 'toto 1 titi',
+			],
+		];
+	}
+
+	/**
+	 * @since 2.7.10 N°5491 - Inconsistent dictionary entries regarding arguments to pass to Dict::Format
+	 * Dict::Format
+	 * @dataProvider FormatProvider
+	 */
+	public function testFormat($sTemplate, $sExpectedTraduction){
+		$sLangCode = \Dict::GetUserLanguage();
+		$aDictByLang = $this->GetNonPublicStaticProperty(\Dict::class, 'm_aData');
+		$sDictKey = 'ITOP::DICT:FORMAT:BROKEN:KEY';
+
+		if (! is_null($sTemplate)){
+			if (array_key_exists($sLangCode, $aDictByLang)){
+				$aDictByLang[$sLangCode][$sDictKey] = $sTemplate;
+			} else {
+				$aDictByLang[$sLangCode] = [$sDictKey => $sTemplate];
+			}
+		}
+
+		$this->SetNonPublicStaticProperty(\Dict::class, 'm_aData', $aDictByLang);
+
+		$this->assertEquals($sExpectedTraduction, \Dict::Format($sDictKey, "1"));
+	}
+}
